refactor(arch): Phase 2 — email → shared/email + entities/email-token, skills → entities/skill
- email dispatcher → shared/email/dispatcher.ts, email types → shared/email/types.ts
- tokenStore → entities/email-token/api.ts, templates → entities/email-token/templates/
- skills loader → entities/skill/api.ts, types → entities/skill/model.ts
- 11개 파일 이동 (소스 6 + 테스트 5), 22개 파일 import 갱신
- infrastructure/email/, infrastructure/skills/ 완전 제거

Phase 2 of FSD migration (PR D: email + skills)
