+++ src/__tests__/entities/email-token/api.test.ts
@@ -13,7 +13,7 @@ import {
     buildEmailTokenKey,
     createEmailTokenStore,
     type EmailTokenValue,
-} from '@/infrastructure/email/tokenStore';
+} from '@/entities/email-token';
 
 const MockRedis = Redis as unknown as jest.Mock;
