+++ src/__tests__/entities/email-token/templates/emailVerificationEmail.test.ts
@@ -1,4 +1,4 @@
-import { buildEmailVerificationEmail } from '@/infrastructure/email/emailVerificationEmail';
+import { buildEmailVerificationEmail } from '@/entities/email-token';
 
 describe('buildEmailVerificationEmail', () => {
     const baseInput = {
