+++ src/app/[symbol]/page.tsx
@@ -9,7 +9,7 @@ import {
 import { buildAssetAboutNode } from '@/domain/seo/assetClassification';
 import { buildDisplayName } from '@/domain/ticker';
 import { getBarsAction } from '@/infrastructure/market/getBarsAction';
-import { countSkillFiles } from '@/infrastructure/skills/loader';
+import { countSkillFiles } from '@/entities/skill';
 import { getAssetInfoCached } from '@/infrastructure/ticker/getAssetInfoCached';
 import { QUERY_KEYS, QUERY_STALE_TIME_MS } from '@/shared/config/queryConfig';
 import {
