+++ src/app/page.tsx
@@ -8,10 +8,7 @@ import { StatsBar, StatsBarSkeleton } from '@/components/home/StatsBar';
 import { TickerCategories } from '@/components/home/TickerCategories';
 import { SymbolSearchPanel } from '@/components/search/SymbolSearchPanel';
 import { JsonLd } from '@/shared/ui/JsonLd';
-import {
-    countSkillFiles,
-    FileSkillsLoader,
-} from '@/infrastructure/skills/loader';
+import { countSkillFiles, FileSkillsLoader } from '@/entities/skill';
 import { SITE_DESCRIPTION, SITE_NAME, SITE_URL } from '@/shared/lib/seo';
 import Link from 'next/link';
 import { cache, Suspense } from 'react';
