+++ src/entities/email-token/api.ts
@@ -1,19 +1,9 @@
 import { Redis } from '@upstash/redis';
 
+export type { EmailDispatcher, EmailMessage } from '@/shared/email/types';
+
 /** Purpose tag namespacing email-token Redis keys so password-reset and email-verification tokens for the same email never collide. */
 export type EmailTokenPurpose = 'password_reset' | 'email_verification';
-
-/** Email message payload passed to an {@link EmailDispatcher} implementation. */
-export interface EmailMessage {
-    /** Recipient email address. */
-    to: string;
-    /** Email subject line. */
-    subject: string;
-    /** HTML body of the email. */
-    html: string;
-    /** Plain-text body of the email. */
-    text: string;
-}
 
 /** Persisted Redis value for an email token: 'pending' (token issued; tokenHash stored for constant-time compare on verify) or 'verified' (email-verification flow only; marker so later registration confirms ownership without re-prompting the code). */
 export type EmailTokenValue =
@@ -46,12 +36,6 @@ export interface EmailTokenStore {
         purpose: EmailTokenPurpose,
         email: string
     ): Promise<EmailTokenValue | null>;
-}
-
-/** 트랜잭셔널 이메일 발송 추상 (Resend/SendGrid/SMTP 등을 use-case에 주입). */
-export interface EmailDispatcher {
-    /** 메시지 발송 — 전송 수락 시 true, 거절 시 false. */
-    sendEmail(message: EmailMessage): Promise<boolean>;
 }
 
 const KEY_PREFIX = 'email_token';
