+++ src/entities/email-token/index.ts
@@ -1,0 +1,15 @@
+export {
+    __resetEmailTokenStoreCacheForTests,
+    buildEmailTokenKey,
+    createEmailTokenStore,
+} from './api';
+export type {
+    EmailDispatcher,
+    EmailMessage,
+    EmailTokenPurpose,
+    EmailTokenStore,
+    EmailTokenValue,
+} from './api';
+export { buildPasswordResetEmail } from './templates/passwordResetEmail';
+export { DEFAULT_SITE_URL } from './templates/passwordResetEmail';
+export { buildEmailVerificationEmail } from './templates/emailVerificationEmail';
