+++ src/entities/email-token/templates/passwordResetEmail.ts
@@ -1,4 +1,4 @@
-import type { EmailMessage } from '@/infrastructure/email/tokenStore';
+import type { EmailMessage } from '@/shared/email/types';
 
 // Duplicates @/lib/seo SITE_NAME/SITE_URL — update both if changed.
 const SITE_NAME = 'Siglens';
