+++ src/entities/skill/index.ts
@@ -1,0 +1,2 @@
+export { countSkillFiles, FileSkillsLoader } from './api';
+export type { SkillsProvider } from './model';
