+++ src/entities/skill/model.ts
