+++ src/infrastructure/auth/requestPasswordResetAction.ts
@@ -2,10 +2,10 @@
 
 import { DrizzleUserRepository } from '@/entities/user';
 import { requestPasswordReset } from '@/infrastructure/auth/use-cases/requestPasswordReset';
-import { createEmailTokenStore } from '@/infrastructure/email/tokenStore';
+import { createEmailTokenStore } from '@/entities/email-token';
 import type { ForgotPasswordFormState } from '@/domain/auth/formTypes';
-import { buildPasswordResetEmail } from '@/infrastructure/email/passwordResetEmail';
-import { createEmailDispatcher } from '@/infrastructure/email/resend';
+import { buildPasswordResetEmail } from '@/entities/email-token';
+import { createEmailDispatcher } from '@/shared/email/dispatcher';
 import { getAuthDatabaseClient } from '@/infrastructure/auth/db';
 
 export async function requestPasswordResetAction(
