+++ src/shared/CLAUDE.md
@@ -12,7 +12,7 @@
 | `shared/ui/` | Primitive UI 컴포넌트: DotSeparator, EyeIcon, InfoTooltip, JsonLd, MarkdownText, tabs/ |
 | `shared/hooks/` | React 의존 일반 hook: useDialog, useEscapeKey, useFocusTrap, useHydrated, useIsMobileViewport 등 |
 | `shared/db/` | Drizzle/Neon client, schema, token encryption, DB config/constants/types |
-| `shared/email/` | (Phase 2에서 추가 예정) Email dispatcher |
+| `shared/email/` | Email dispatcher (Resend/Noop) + email types (EmailMessage, EmailDispatcher) |
 | `shared/cache/` | (Phase 2에서 추가 예정) Redis client |
 | `shared/api/` | (Phase 1 PR B에서 추가 예정) HTTP client, third-party API wrapper |
